requirements.txt
llama-index-core==0.14.3
llama-index-vector-stores-faiss==0.5.1
llama-index-llms-openai==0.5.6
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai>=0.3,<1

sentence-transformers==3.0.1

faiss-cpu>=1.9.0,<1.13

ocrmypdf>=16,<17
pymupdf>=1.24,<1.27
pillow>=11.1.0,<12

streamlit>=1.39.0,<1.41.0

openai>=1,<3
python-dotenv==1.0.1
requests==2.32.3